Circle on the end of the curve. Preven unnecessary glUseProgram() calls. Remove unused parts of textureSurface.vert. Prevent resizing curve buffer to 0.

+++ AndroidApp/app/src/main/jni/DrawableObject.cpp
@@ -1,6 +1,6 @@
 #include "DrawableObject.h"
 
-void DrawableObject::setZOrder(int z){
+void DrawableObject::setZOrder(float z){
     zCoordinate=0.001f*z;
 }
